task(refactor): Converts single command to multi
