+++ packages/core/package.json
@@ -13,8 +13,8 @@
   "devDependencies": {
     "@types/chai": "^4",
     "@types/jest": "^25.1.3",
-    "@types/node": "^13",
     "@types/js-yaml": "^3.12.2",
+    "@types/node": "^13",
     "@types/resolve": "^1.14.0",
     "chai": "^4.2.0",
     "cosmiconfig": "^6.0.0",
